Stat cleanup.
From the mailing list:

Ok, first thing: clean up the help text. I realize what's there is copied verbatim from the man page, but that man page sucks. ("modification time" vs "change time"?) Took a bit of finagling to fit it in 80x24, but just made it.

GLOBALS() indent was still tab, change to two spaces. And I tend to put a blank line between options lib/args.c automatically fills out and normal globals.

We never do anything with date_stat_format() but immediately print it, might as well make the function do it.

The types[] array in do_stat() is a rough edge. Hmmm... there's no else case that sets the type in case it was unknown (such as 0). In theory, this never happens. In practice it means I can cheat slightly, given this observation:

  $ find linux -name stat.h | xargs grep 'S_IF[A-Z]*[ \t]'
  linux/include/uapi/linux/stat.h:#define S_IFMT  00170000
  linux/include/uapi/linux/stat.h:#define S_IFSOCK 0140000
  linux/include/uapi/linux/stat.h:#define S_IFLNK  0120000
  linux/include/uapi/linux/stat.h:#define S_IFREG  0100000
  linux/include/uapi/linux/stat.h:#define S_IFBLK  0060000
  linux/include/uapi/linux/stat.h:#define S_IFDIR  0040000
  linux/include/uapi/linux/stat.h:#define S_IFCHR  0020000
  linux/include/uapi/linux/stat.h:#define S_IFIFO  0010000

I.E. the only place the I_IFBLAH constants occur a stat.h header in current linux code is in the generic stuff, it doesn't vary per target. (The access permission bits are actually subtly standardized in posix due to the command line arguments to chmod, although I'm sure cygwin finds a way to break. But the type fields, not so much. But linux has to be binary compatible with itself foreverish, and that's all I really care about.)

So, we have ALMOST have this going by twos, except there's no 8 and there is a 1. so let's make the 1 the default, feed a blank string into the 8... No, duh: octal. So it's actually 2, 4, 6, 8, 10, 12. So make the loop look like:

  filetype = statf->st_mode & S_IFMT;
  TT.ftname = types;
  for (i = 1; filetype != (i*8192) && i < 7; i++)
    TT.ftname += strlen(TT.ftname)+1;

Yes that's linux-specific, and I think I'm ok with that.

Printing all zeroes and pretending that's nanosecond resolution... either support it or don't. Let's see, supporting it is stat->st_atim.tv_nsec and similar... no mention of nanoseconds in strftime() (et tu, posix2008?) so pass it as a second argument and append it by hand... (Need to test that against musl...)

When we hit an unknown type in print_it() we print the literal character, which is right for %% but what about an unknown option?

  $ stat -c %q /
  ?

Eh, I guess that's a "don't care". It didn't die with an error, that's the important thing.

I have a horrible idea for compressing the switch/case blocks, but should probably check this in and get some sleep for right now...