xzcat: remove XZ_(PREALLOC|SINGLE), inline xz_dec_bcj_create
Because we only use XZ_DYNALLOC, there's a bunch of dead code.
This patch removes the #ifdef's and if()s associated with support for
multiple modes.
single_call was only used to store the mode; it is no longer needed.
A little bit of reorganization was needed to reduce the number of prototypes.
Documentation associated with dead code was dropped.

There are still some relics of multiple modes in the continued presence
of "XZ_DYNALLOC" and xz_mode.

Additionally, I inlined xz_dec_bcj_create; it was called once.

This loses about 125 lines, mostly comments.