Minor cosmetic tweaks to expand.
Working my way through the to-review list that predates the "pending"
directory. This gets expand off my to-review list. (Proof that "need to
review" doesn't mean it's in bad shape, this command was fine. Changed
capitalization in the help text because I'm trying to have "user supplies
this value" be all caps, switched a read() to readall() although I'm not
sure modern kernels actually allow -EINTR to generate zero length reads
anymore, and since most of the loopfiles() target functions are called
do_commandname() changed the name to that just so it's regular. None of
the changes are actually important. :)