Fix use-after-free spotted by Ashwini Sharma's static analysis.
We xstrdup() an optargs string to avoid modifying our environment space
(because it can change what "ps" shows to other processes), and then parse
out colon delimited strings and save them in globals that can later be used
in the -v codepath and so on. But those globals _aren't_ strdup (no point)
which means we can't free the string while we're still using pointers into
the middle of it. So move the free to the end.

(I hardly ever test with CFG_TOYBOX_FREE switched on because even nommu
doesn't need it.)