Variant of a patch from Ashwini Sharma, making df /dev/node work and tweaking the spacing.
I didn't apply the POSIXLY_CORRECT gnuism because it's a can of worms (as would
be LSB_CORRECT), and you can presumably alias df="df -P" if you want that.

Possibly in future I should factor out the "readahead and align columns based
on measuring the largest value in each" code from ls and apply it here.