Fix more with missing files.
Previously we'd go into an infinite loop because we weren't
incrementing optargs.

Also add a missing flush so an error on stderr won't overtake the
escape code that resets reverse video.

Disclaimer: the new behavior isn't exactly like the desktop version;
surprisingly they try to open the next file _before_ they prompt. That
feels weird to me as a user, and seems like it would lead to a more
awkward implementation, but if you're more concerned about
authenticity...