Fix touch(1) date parsing on tzcode-derived systems.
The IANA tzcode implementation of mktime (used on Android and BSDs) sets
errno in some cases where it doesn't return -1 to indicate failure, so the
existing test always failed on those systems.

I don't think glibc ever sets errno (which is fine by ISO C, but not POSIX).

Other uses of mktime in toybox are already fine. This one would have been
caught by the existing tests if I was running them on the device :-(