Updated Android roadmap.
Since we're actually getting close to the end of the toolbox list,
expand the "other stuff" with everything that's currently in /system/bin
that could reasonably be replaced by toybox (and remove things that might
sound practical, but actually aren't). Also update the list of things
being built from pending.