Fix hang in losetup test.
Switch the printf over to %s because the input is actually hex (so %d
is wrong), but without the leading "0x" that %x would require.

For some reason the NODE assignment wasn't providing any input to awk
(hence the hang), and awk wasn't looking for the correct field anyway.

The tests still fail for me:

  losetup: /dev/block/loop0: No such device or address
  losetup: /dev/block/loop0: No such device or address
  losetup: file: No such file or directory
  losetup: needs 1 arg (see "losetup --help")

But at least now they allow the other tests to continue!