getconf: fix glibc NPROCESSORS_ONLN.
glibc doesn't have _XOPEN_UUCP (though bionic does), which meant that the
generated array of values was out of sync with the hand-written array of
names.

This patch removes that by using a unified array and the preprocessor.

A side benefit of this for me is that it makes toybox easier to integrate
in the AOSP build system (the less shell script magic, the better).