macOS: move getmountlist.c functions into portability.c.
I think this was the preferred option for this code that turns out
to be very different on Linux vs BSD. I don't yet have a BSD
implementation, and I'm not likely to have chance to work on one
in the near future, so this just #ifdefs it out for macOS right
now.

With this (and a suitably minimal .config), toybox builds for macOS
out of the box.