Fix UDP checksum verification error.
From RFC 768, if UDP packet checksum computation yields a result of
zero, change it to hex 0xFFFF. The current udhcpc checksum verification
would yield false positive for this case. A better way is to compute the
checksum with the original checksum field and the result should be zero
for good udp packet.

Signed-off-by: Yangchun Fu <yangchun@google.com>