hostname: fix behavior when in jail.
Only -d and -f should cause a DNS lookup. The rest should just act
directly on the result of gethostname(3). Encountered with the AOSP
buildbots' use of nsjail, but tested with both the Debian hostname and
toybox hostname thus:

```
unshare -Uunr sh
hostname android-build
hostname
hostname -s
hostname -d
hostname -f
```

(Not sure how to add that to the tests.)

Also fix a SEGV with -s if the hostname doesn't contain a '.'.

Also switch to the FLAG() macro.

Also add the missing -s to the synopsis.

Bug: http://b/123123255